% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_func.R
\name{conduct_lmm_int_constant}
\alias{conduct_lmm_int_constant}
\title{Conduct analysis using lmerTest::lmer()}
\usage{
conduct_lmm_int_constant(..., data, predictors, outcome,
  constants = NULL, ranef = "ther_id", ranslope = "1")
}
\arguments{
\item{...}{Additional variables}

\item{data}{Data to use}

\item{predictors}{List of predictors to include in formula}

\item{outcome}{Variable to use as the outcome of the lm() call}

\item{constants}{List of variables that will be in EVERY permutation.
Entered separately from predictors at present.}

\item{ranef}{The random effects variables}

\item{ranslope}{Variables (if any) to be included as slopes.}
}
\value{
A model of type lmerMod
}
\description{
Conduct analysis using lmerTest::lmer()
}
\details{
This function only tolerates one grouping variable at present.
Or more directly, one parenthetical feature per formula, so it is possible that a
specification like \code{(1 | therapist/patient)} could work?
It should tolerate multiple slope variables, though.
}
\examples{
out <- conduct_lmm_int_constant(data = test_data,
  predictors = c("n_tx_eval_tot", "n_noshow_tot", "week_since_first_attend", "ther_id"),
  constants = c("first_BDI"),
  outcome = "BDI")

}
