% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_func.R
\name{conduct_prepost}
\alias{conduct_prepost}
\title{Conduct analysis of pre-post data WITHOUT constants.
Obsolete, use conduct_prepost_constant() instead.}
\usage{
conduct_prepost(..., data, predictors, outcome)
}
\arguments{
\item{...}{Additional variables}

\item{data}{Data to use}

\item{predictors}{List of predictors to include in formula}

\item{outcome}{Variable to use as the outcome of the lm() call}
}
\value{
an lm() model
}
\description{
Conduct analysis of pre-post data WITHOUT constants.
Obsolete, use conduct_prepost_constant() instead.
}
\examples{
outcome <- conduct_prepost(data = test_data,
  predictors = c("n_tx_eval_tot", "n_noshow_tot", "week_since_first_attend", "ther_id"),
  outcome = "BDI")
}
