% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_func.R
\name{conduct_prepost_constant}
\alias{conduct_prepost_constant}
\title{Conduct analysis of pre-post data WITH constants.}
\usage{
conduct_prepost_constant(..., data, predictors, outcome,
  constants = NULL)
}
\arguments{
\item{...}{Additional variables}

\item{data}{Data to use}

\item{predictors}{List of predictors to include in formula}

\item{outcome}{Variable to use as the outcome of the lm() call}

\item{constants}{List of variables that will be in EVERY permutation.
Entered separately from predictors at present.}
}
\value{
a model of type lm
}
\description{
Conduct analysis of pre-post data WITH constants.
}
\examples{
outcome <- conduct_prepost_constant(data = test_data,
  predictors = c("n_tx_eval_tot", "n_noshow_tot", "week_since_first_attend", "ther_id"),
  constants = c("first_BDI"),
  outcome = "BDI")
}
