% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_func.R
\name{extract_target_lm}
\alias{extract_target_lm}
\title{Extract values of interest from lm and lmerMod objects}
\usage{
extract_target_lm(..., model, target)
}
\arguments{
\item{...}{Additional variables}

\item{model}{Model to extract from}

\item{target}{Name of the variable predictor of interest}
}
\value{
a 1x6 data frame with Estimate, SE, t val, and p val, lower, and upper
}
\description{
Extract values of interest from lm and lmerMod objects
}
\examples{
extract_target_lm(model = model, target = "age")
}
