% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_func.R
\name{isolate_predictors}
\alias{isolate_predictors}
\title{Isolate one permutation of non-NA predictors from a previously-defined
row of a matrix of predictor permutations}
\usage{
isolate_predictors(..., x)
}
\arguments{
\item{...}{Additional arguments passed on from other functions}

\item{x}{The predictor permutation to use}
}
\value{
A vector with just the non-NA predictors
}
\description{
Isolate one permutation of non-NA predictors from a previously-defined
row of a matrix of predictor permutations
}
\details{
This function exists only to remove NA values from lists of predictors
If x is entirely NA, returns NULL, which is compatible with lm() and lmer() formula definitions.
}
\examples{
isolate_predictors(x = c("age", NA, "first_BDI"))

}
