% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_func.R
\name{make_predictors}
\alias{make_predictors}
\title{Generate permutations of predictors with varying options}
\usage{
make_predictors(..., options, target)
}
\arguments{
\item{...}{Additional arguments}

\item{options}{A list of variables in the data set}

\item{target}{The variable in the data set that will be analyzed for its effect}
}
\value{
A matrix in which each row is a unique permutation of the options for each variable
}
\description{
Generate permutations of predictors with varying options
}
\examples{
make_predictors(options = c("BAI", "first_BDI"), target = "age")
}
