% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_func.R
\name{run_data}
\alias{run_data}
\title{Run a single data permutation through multiple predictor permutations}
\usage{
run_data(..., data, predictors, target = predictors$target[1],
  outcome = "BDI", constants = "first_BDI")
}
\arguments{
\item{constants}{}
}
\description{
Run a single data permutation through multiple predictor permutations
}
