% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_func.R
\name{run_data_lmm}
\alias{run_data_lmm}
\title{Run a single data frame through multiple predictor permutations with lmer()}
\usage{
run_data_lmm(..., data, predictors, target = predictors$target[1],
  outcome = "BDI", constants = "first_BDI", ranef = "ther_id",
  ranslope = "1")
}
\arguments{
\item{ranslope}{}
}
\description{
Run a single data frame through multiple predictor permutations with lmer()
}
