% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_func.R
\name{run_multiple_trims}
\alias{run_multiple_trims}
\title{Run a complete specification curve analysis}
\usage{
run_multiple_trims(..., data, min_ses_options = NA,
  max_ses_options = NA, max_courses_options = NA,
  min_sev_BDI_options = NA, min_sev_BAI_options = NA,
  min_obs_options = NA, predictors, target, outcome = "BDI",
  constants = "first_BDI", include_ran_thers = TRUE)
}
\arguments{
\item{...}{Additional arguments}

\item{data}{Data set}

\item{min_ses_options}{Minimum number of sessions attended}

\item{max_ses_options}{Maximum number of sessions attended}

\item{max_courses_options}{Maximum number of courses allowed}

\item{min_sev_BDI_options}{Minimim initial severity of BDI}

\item{min_sev_BAI_options}{Minimum initial severity of BAI}

\item{min_obs_options}{Minimum number of observations per patient}

\item{predictors}{List of predictors to permute}

\item{target}{The target effect}

\item{outcome}{The outcome variable}

\item{constants}{List of variables to including in all permutations}

\item{include_ran_thers}{Logical: should random therapists be included as an option?}
}
\value{
a \code{tibble} with outcome of all analyses
}
\description{
Given a data frame and the options, trim it in every permutation
and run each with every predictor permutation available.
}
