% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{spec_curve}
\alias{spec_curve}
\title{Plot the main specification curve}
\usage{
spec_curve(data, y = data$Estimate, lower = data$lower,
  upper = data$upper, title = NULL)
}
\arguments{
\item{data}{Data to use}

\item{y}{y variable, defaults to data$Estimate}

\item{lower}{vector representing lower bounds of CIs}

\item{upper}{vector representing upper bounds of CIs}

\item{title}{title to pass to ggtitle()}
}
\value{
a ggplot plot object
}
\description{
Plot the main specification curve
}
\examples{
spec_curve(sim_out)
}
