% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_func.R
\name{trim_data}
\alias{trim_data}
\title{Trim a data set based on one permutation of variabls}
\usage{
trim_data(..., data, min_ses_num = 0, max_ses_num = 10000,
  max_courses = 1, completers_only = FALSE, min_sev_baseline = 0,
  min_sev_BDI = 0, min_sev_BAI = 0, min_obs = 0, use_obs = c("all",
  "prepost", "post"), param_list = NULL)
}
\arguments{
\item{...}{additional variables passed on from other functions}

\item{data}{data to use}

\item{min_ses_num}{Minimum number of sessions required}

\item{max_ses_num}{Maximum number of sessions allowed}

\item{max_courses}{Maximum number of courses of treatment allowed}

\item{completers_only}{Logical. Should only completers be used?}

\item{min_sev_baseline}{Minimum initial severity, outcome not defined.}

\item{min_sev_BDI}{Minimum initial severity on the BDI.}

\item{min_sev_BAI}{Minimum initial severity on the BAI.}

\item{min_obs}{Minimum number of observations per patient.}

\item{use_obs}{Which observations to use? Options: "prepost", "post", and "all".}

\item{param_list}{Optional list of parameter values that overwrite existing parameters if included.}
}
\value{
a trimmed dataset
}
\description{
Trim a data set based on one permutation of variabls
}
\examples{
trim_data(data = sim_data)
}
