% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificationCurves-package.R
\docType{package}
\name{specificationCurves-package}
\alias{specificationCurves}
\alias{specificationCurves-package}
\title{specificationCurves: Specification Curves}
\description{
Some functions for making specification curves.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewmcaleavey/specificationCurves}
  \item Report bugs at \url{https://github.com/andrewmcaleavey/specificationCurves/issues}
}

}
\author{
\strong{Maintainer}: Andrew McAleavey \email{andrew.mcaleavey@gmail.com} (0000-0001-5986-2033)

}
\keyword{internal}
